% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_badge.R
\name{coverage_percentage}
\alias{coverage_percentage}
\title{Calculate the coverage percentage from cobertura.xml}
\usage{
coverage_percentage(cobertura_path = "cobertura.xml", delete_xml = FALSE)
}
\arguments{
\item{cobertura_path}{Path to the `cobertura.xml` file. (String)}

\item{delete_xml}{If TRUE deletes the `cobertura.xml` after execution. (Logical)}
}
\description{
Calculate the coverage percentage from cobertura.xml
}
