% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_badge.R
\name{create_coverage_badge}
\alias{create_coverage_badge}
\title{Generates a Coverage Badge in SVG}
\usage{
create_coverage_badge(coverage_perc, path = "coverage_badge.svg")
}
\arguments{
\item{coverage_perc}{Percentage reported by cobertura::coverage_percentage (Numeric)}

\item{path}{Path in with the SVG file should be written. (String)}
}
\description{
Generates a Coverage Badge in SVG
}
