% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_badge.R
\name{.coverage_url}
\alias{.coverage_url}
\title{Create a Shields.IO URL to download a SVG}
\usage{
.coverage_url(coverage_perc)
}
\arguments{
\item{coverage_perc}{Percentage (Numeric)}
}
\description{
Create a Shields.IO URL to download a SVG
}
