% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_badge.R
\name{generate_cobertura}
\alias{generate_cobertura}
\title{Generates a cobertura.xml file in the base directory}
\usage{
generate_cobertura()
}
\description{
Generates a cobertura.xml file in the base directory
}
